requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pdfplumber>=0.9.0
PyMuPDF==1.23.8
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
python-dotenv==1.0.0
numpy==1.24.3
openpyxl>=3.1.0
xlrd==2.0.1
psutil>=5.9.0
urllib3>=2.0.0 